+++ app/javascript/app/components/sectors-agriculture/drivers-of-emissions/historical-emissions-graph/historical-emissions-graph-styles.scss
@@ -15,10 +15,10 @@
   margin-bottom: 40px;
 }
 
-.dropdown {
-  max-width: 300px;
-}
-
 .filtersGroup {
   display: flex;
+  & > * {
+    margin-right: 30px;
+    max-width: 220px;
+  }
 }
